requirements.txt
protobuf==6.30.2
pyproj==3.7.1
python-dotenv==1.1.0
Requests==2.32.3
Shapely==2.1.0
